requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlmodel>=0.0.8
python-dotenv>=1.0.0
pandas>=2.0.3
openpyxl>=3.1.2
chromadb>=0.4.13
llama-index>=0.8.45
llama-index-vector-stores-chroma>=0.1.1
llama-index-embeddings-openai>=0.1.1
python-multipart>=0.0.6
requests>=2.31.0